requirements.txt
RPi.GPIO
pytz
influxdb

